requirements.txt
asgiref==3.4.1
boto3==1.20.34
botocore==1.23.34
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.4.1
Django==3.2.5
django-cors-headers==3.11.0
django-cors-middleware==1.5.0
django-extensions==3.1.5
django-filter==21.1
django-hosts==5.0
django-utils-six==2.0
djangorestframework==3.13.1
docutils==0.13.1
gunicorn==21.2.0
idna==3.3
itsdangerous==0.24
jmespath==0.9.1
mysqlclient==2.1.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
